+++ src/Form/ConfirmEmailType.php
@@ -5,13 +5,15 @@ namespace App\Form;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Security;
 
 class ConfirmEmailType extends ModalFormType
 {
-    public $security;
+    private $security;
 
-    public function __construct(RequestStack $requestStack, Security $security)
+    public function __construct(RequestStack $requestStack, Security $security, UrlGeneratorInterface $urlGenerator)
     {
         parent::__construct($requestStack);
 
@@ -26,10 +28,19 @@ class ConfirmEmailType extends ModalFormType
             ->setMethod('POST')
 
             ->add('email', EmailType::class, [
-                'data' => $user ? $user->getEmail() : null,
+                'data' => $user ? $user->getEmail() : $options['email'],
                 'label' => 'Email',
                 'help' => 'Your email',
             ])
         ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        parent::configureOptions($resolver);
+
+        $resolver->setDefaults([
+            'email' => null,
+        ]);
     }
 }